# [Core] Introduce popleft_n and append_n in FreeKVCacheBlockQueue to further optimize block_pool