--- a/vllm/v1/core/block_pool.py
+++ b/vllm/v1/core/block_pool.py
@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+import itertools
 from collections import defaultdict
 from collections.abc import Iterable
 from typing import Callable, Optional
@@ -215,13 +216,14 @@ class BlockPool:
                 f"Cannot get {num_blocks} free blocks from the pool")
 
         ret: list[KVCacheBlock] = self.free_block_queue.popleft_n(num_blocks)
-        for block in ret:
-            assert block.ref_cnt == 0
-            block.ref_cnt += 1
 
         if self.enable_caching:
             for block in ret:
                 self._maybe_evict_cached_block(block)
+
+        for block in ret:
+            assert block.ref_cnt == 0
+            block.ref_cnt += 1
 
         return ret
 
@@ -278,12 +280,12 @@ class BlockPool:
             ordered_blocks: A list of blocks to free ordered by their eviction
                 priority.
         """
-        for block in ordered_blocks:
+        # Create 2 iterators to allow iterateing over ordered_blocks twice.
+        blocks_iter1, blocks_iter2 = itertools.tee(ordered_blocks)
+        for block in blocks_iter1:
             block.ref_cnt -= 1
-
-        # null_block should not be added to the free list.
         self.free_block_queue.append_n([
-            block for block in ordered_blocks
+            block for block in blocks_iter2
             if block.ref_cnt == 0 and not block.is_null
         ])