@Jialin 2025-07-19
resolve https://github.com/vllm-project/vllm/issues/21141